Corrigido limite da variável do número de ações, intermodulações e pinos para 100. Adicionado função para verificar se o número de ações, intermodulações e pinos é menor que o limite antes de adicionar um novo item. Atualizado a documentação para refletir as mudanças.

+++ src/mqtt_manager.cpp
@@ -113,7 +113,7 @@ void fV_mqttCallback(char* topic, byte* payload, unsigned int length) {
         int startIdx = expectedPrefix.length();
         int endIdx = topicStr.lastIndexOf("/set");
         String pinNumberStr = topicStr.substring(startIdx, endIdx);
-        uint8_t pinNumber = pinNumberStr.toInt();
+        uint16_t pinNumber = pinNumberStr.toInt();
         
         // Encontrar índice do pino
         uint8_t pinIndex = fU8_findPinIndex(pinNumber);
